// polytrader-console/src/lib/api/models/live-strategies-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Polytrader Control API
 * Control plane API for Polytrader platform
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Active live strategies response.
 */
export interface LiveStrategiesResponse {
    /**
     * List of active strategy IDs
     */
    'active_strategies': Array<string>;
}